fix(composer): Fixed composer not picking dev version.
