+++ blackduck-artifactory-common/src/main/java/com/synopsys/integration/blackduck/artifactory/modules/inspection/externalid/composer/model/VersionSource.java
@@ -11,7 +11,7 @@ import com.google.gson.annotations.SerializedName;
 
 public class VersionSource {
     @SerializedName("reference")
-    private String reference;
+    public String reference;
 
     public String getReference() {
         return reference;
